Use system server jar module stem instead of its name
In the case a system server jar is added as a prebuilt
its Name() will prepend prebuilt_ to the module name.

By itself this doesn't cause any issue, but the format
of PRODUCT_SYSTEM_SERVER_JARS expects the second part
to match the "installed" jar and depending on that it
fills the AllPlatformSystemServerJars, that now can't be
indexed by the Name() of the prebuilt module.

To solve the issue use the module stem as GetSystemServerDexLocation
will construct a path based on that.

Test: m, prebuilt felix-services passes dexpreopt check
Change-Id: I7e61b958bb1772109446e56bad7a51dd81e8659e